bot/requirements.txt
pyTelegramBotAPI== 4.3.1
mysql-connector==2.2.9
mysqlclient==2.1.0